+++ tools/widl/hash.h
@@ -28,6 +28,6 @@ typedef enum tag_syskind_t {
     SYS_MAC
 } syskind_t;
 
-extern unsigned long lhash_val_of_name_sys( syskind_t skind, LCID lcid, LPCSTR lpStr);
+extern unsigned int lhash_val_of_name_sys( syskind_t skind, LCID lcid, LPCSTR lpStr);
 
 #endif
